requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
openpyxl==3.1.5
pillow==10.2.0
PyJWT==1.7.1
pytz==2024.1
pywin32==306
requests==2.31.0
slack_sdk==3.29.0
slackclient==2.9.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
